fix: Rename Components That Only Need One Word
By default, Vue 3 wants you to name components with two words, but we
have some components that only need to be one word, so we change the
`eslintrc.js` to allow us to name components with one word and we
rename components.
