+++ src/views/UserManual.vue
