+++ src/views/__tests__/HomeView.spec.ts
@@ -1,8 +1,8 @@
-import HomeView from '../HomeView.vue'
+import Home from '../Home.vue'
 import {expect, test} from 'vitest'
 import {mount} from '@vue/test-utils'
 
 test('should contain link to scope', () => {
-    const wrapper = mount(HomeView, {shallow: true})
+    const wrapper = mount(Home, {shallow: true})
     expect(wrapper.html()).toContain('/scope')
 })
