+++ src/views/minor/Footer.vue
